FINALLY fixed the problem that prevented training - see description
the input to the model through the predict() must be a list [a, b, c] of arrays, where each array a, b, and c has the dimensions (samples, dim, dim, dim, ..., features)

when you tell the model what input shape it will recieve, it is the shape of the arrays a, b, or c but without the (sample, ) dimension

so i.e. the (20, 20, 1) board needed to be placed within an array of size one, same with the velocity and position tensors, and then join those arrays of length 1 into a list [] to pass as a parameter!